Calibrating monitor and testing switch

+++ slot-in-smart-meter/slot-in-smart-meter/communication.h
@@ -17,7 +17,7 @@ public:
 
 	int connectToNetwork();
 	int connectToServer();
-	int sendData(int);
+	byte sendData(float);
